fix(text): Fix a crashing issue with link highlighting under Ventura.

+++ Text Handling/RDTextView.m
@@ -138,7 +138,7 @@
 	NSRange			charRange;
 	NSRange			linkRange;
 	NSRectArray		linkRects;
-	unsigned		linkCount;
+	unsigned long	linkCount;
 	int				scanLoc;
 	int				index;
